project-expense-tracking-system/requirements.txt
fastapi==0.115.12
pydantic==2.11.4
mysql-connector-python==9.3.0
openpyxl==3.1.5
streamlit==1.45.0
requests==2.32.3
pandas==2.2.3
pytest==8.3.5